[REFACTORING] Split IMAPServerCondStoreTest
